Removing usage of legacy driving animations
This commit uses the modern driving animation builder for the
initial driving segment at the start of a flythrough. This removes
the final remaining usage of the old legacy driving animation
logic, so it is removed as well.

However, this requires updating the path building logic for the
modern driving animation builder, because this breaks the previous
assumption that the initial points in a driving path are always
one road segment apart. (This is probably a good thing anyway
because it removes a previously hidden assumption in the code,
which could have caused confusion in the future).