Better function/argument names
`determineAzimuthAngle` seemed confusing because it was vague. The
new function name is still maybe also a little vague, but it at
least indicates a little more that this function returns a
_target_ angle to rotate to.

`oldTargetPositionX` and `oldTargetPositionZ` seemed like
confusing names. These names seemed to refer to incidental facts
about how their values are used externally, which isn't relevant
to how they are used in this function. The new names hopefully
more clearly indicate that they refer to a starting point from
which a rotation is performed.