No camera jump between consecutive driving paths
Previously, if the camera angle at the end of one driving path was
different from the starting camera angle of the next path, the
camera would suddenly jump to the new angle when the new path
began. This was jarring and disorienting.

Now the camera will smoothly rotate to the new angle before
beginning the new driving path, similar to how a U-turn works. The
turn is different from a normal driving turn and not as nice since
it occurs in place (which is not how a real car turns), but it at
least prevents a sudden camera jump.