Neighborhood site selection smarter about water
Previously, when determining how suitably flat terrain was for a
neighborhood to be placed, the presence of water was not taken
into account (except for disallowing neighborhood centers on
points which contained water). However, this meant that a
neighborhood could be centered around a point of land that was
mostly surrounded by water. Also, land that was underwater but
flat would count positively towards placing a neighborhood, even
though that land couldn't actually be built on.

This commit prevents underwater land from counting towards a
potential neighborhood center's flatness score. It also
disqualifies a point from being a neighborhood center if it is
surrounded by too much water. Currently "too much" is defined as
half or more of the terrain in the flatness window being water.
It's possible there's a better value for "too much", but this at
least seems to prevent the worst cases of neighborhoods without
enough land to build on.