Removing redundant calculation
As configured, `MAX_HILLINESS` did not do anything. Because it was
used as a divisor, the value of `1.0` always resulted in the
quotient being the same as the dividend. I'm not sure what a
sensible different value would be, or what this was originally
intended to do in the first place (the message for the commit it
was added in, f4ca52df5d347240e59b0d043343ef057e94dfba, didn't
clarify things), so I'm removing it altogether. It can always be
re-added in the future if desired.