Wider Y rotations allowed during flythrough
Previously, when manually rotating the camera during a flythrough,
the camera could only be rotated around the Y axis as far as 90
degrees from the center point. Now it can be rotated as far as 180
degrees.

In practical terms, this means that while previously you could
rotate as far as due left or due right from the vehicle's straight
ahead orientation, you can now rotate as far as looking fully
backwards. After trying this out it seems like an improvement,
because it removes cases where you want to look at something, but
the camera won't rotate far enough to see it.

This doesn't change the rate that the camera rotates while the
mouse/touch is moved - it will still rotate 90 degrees for a drag
halfway across the window. The difference is that now instead of
stopping after this drag distance is reached, the rotation can be
continued across the other half of the window as well.