Adding experimental alternate birdseye animation
The new animation flies between neighborhood centers using a
smooth curved path (specifically, a Catmull-Rom spline). This
seems like a more interesting flying animation that also more
closely approximates a helicopter. The current animation, which
only moves in the cardinal directions, seems stilted in
comparison.

At the moment this is only a prototype/proof of concept which
needs more work before it can replace the current animation. In
particular, some turns are too sharp and disorienting, and there
is not a smooth transition from the driving animation.