Better function interface for `BridgeBuilder.buildBridge()`
Previously, this function was given a starting point and an ending
point to build a bridge between. However, the ending point was
really a _candidate_ ending point, not necessarily where the
bridge would actually end. Also, the ending point had to be
exactly 1 unit away from the starting point in one of the cardinal
directions for the bridge building algorithm to work, which was
potentially confusing.

In the new function interface, a starting point and a target
_direction_ are given instead. This seems easier to understand,
and also allows simplifying some calling code as well.