Adding validation that bridge direction is valid
Bridges (like all roads) must be built in one of the cardinal
directions. This commit causes an error to be thrown if an attempt
is made to do otherwise. In theory this error should never be
thrown as long as calling code is correct, but if not it will
loudly alert the issue.