Preventing large jumps at end of flythrough animation
Since "flythrough mode" allows tilt angles that are more
horizontal than are allowed in "interactive map mode", the camera
sometimes needs to animate to the minimum "interactive map mode"
tilt angle at the end of a flythrough. Previously this sometimes
caused the camera to tilt away from its position by an excessive
amount, resulting in the camera ending up far away from its
original position. This was caused by using the world position at
the center of the browser window to determine the center of tilt,
which sometimes resulted in the center of tilt being a far off
position in the "sky".

This commit fixes the animation to ensure that the center of tilt
for the transition animation is always a constant distance away
from the current camera position, regardless of the terrain or
camera position/rotation.

As a side effect this also partially fixes a bug in which looking
up at a skyward angle at the end of flythrough would result in the
camera being locked in an upward facing position, with no obvious
way to fix other than restarting a flythrough. This is a partial
fix because although the camera will now animate back to a normal
position in this situation, the animation will cause the camera to
go underground and animate back up to the correct position. Still,
this is an improvement because that seems like more of a
"cosmetic" problem, rather than something that essentially
prevented the app from functioning properly.