Camera tilt gesture prevented from going under ground
Previously, unlike the pan, zoom, and azimuth rotation gestures,
there were no restrictions in place which prevented the tilt
gesture from going under the terrain. When tilting backwards into
steep terrain, it was sometimes possible for the camera to go
under the ground towards the end of the gesture.

Now the camera will stop moving once a tilt gesture places the
camera closer to the ground that the required minimum distance.
When using the "Tilt" slider in the UI, the slider value will be
allowed to go the full range, but will not affect the camera once
it would cause the camera to go under the ground.