Preventing issues when "Stop Tour" button double-clicked
Previously, clicking the "Stop Tour" button more than once in the
brief interval between when a flythrough stop is requested and
when it actually stops could cause problems. Specifically, the
button could be labeled "Stop Tour" even after the flythrough was
stopped, and the camera could also briefly "bounce". But thanks to
this commit, no longer.