Zoom in/out buttons properly synced to frame count
Previously, when the zoom in/out buttons were pressed, only 1
frame of the zoom animation would be processed per timer tick,
regardless of how many frames were scheduled to be processed. This
meant that the zoom animation would not occur at the proper rate
unless the timer scheduled exactly 1 frame per tick.

Now the animation will properly advance the correct number of
frames, similar to the other animations that exist.