Easier to understand control flow
This new code seems easier to understand, because it puts the
"stop flythrough" logic for `TimerLoop` in a single place (i.e.
the `stopFlythough()` function). It also seems a little cleaner to
put the internal `TimerLoop` state in the correct state before
broadcasting to the outside world (although in practice that
shouldn't have any practical effect).