Removing no-op statement
Since `syncToCamera()` only does something if the mode is not the
"interactive" mode, but `reset()` is only called in "interactive"
mode (when a new city is created), this call to `syncToCamera()`
was effectively a no-op. In addition, `syncToCamera()` should
ideally only be called just before re-rendering the scene in the
timer tick function, so that is another reason to remove this
call.