Fixing transition animation when camera angle overridden
Previously, the transition animation from "flythrough mode" to
"interactive mode" was not correct when the "vehicle" was in
driving mode and the camera was manually locked in a different
direction from the vehicle's forward direction. If the camera was
tilted upward, the camera would briefly travel under the ground to
reach the final target position.

Now the camera should correctly transition without travelling
underground. As a side effect, `MapCameraTiltAnimation` can be
removed, because it is no longer used.