Immediate switch to "interactive" mode when possible
Previously, when a flythrough was stopped and the camera already
had an X-axis rotation that was valid for the "interactive" camera
mode, there was a brief delay before switching to that mode. In
addition, if the camera was also manually locked in a different
angle from the vehicle's forward looking direction, the camera
briefly rotated even though such a rotation shouldn't be
necessary.

Now, in this scenario the flythrough will be stopped immediately,
and the camera will not perform any unnecessary rotations.