Changing mode name from "interactive" to "manual"
"Interactive" is a possibly confusing name for this mode in the
current app, since it's also possible to interactively rotate the
camera around during a flythrough. "Manual" is intended to reflect
that in this mode you are in charge of manually moving the camera
around the city, instead of relying on the flythrough doing that
for you.