"Rotation"/"Tilt" sliders reset after city generation
Previously the values for these sliders weren't reset to reflect
the default camera position/rotation after a new city is
generated. The values would correct themselves as soon as the city
was interacted with, but if the first interaction was using one of
the sliders the camera could jump if the out-of-sync slider values
were far away from the camera's actual position/rotation.