Better behavior when no valid neighborhood site exists
Previously if a valid neighborhood site could not be selected (for
example because the terrain was heavily flooded with water), the
top left corner of the terrain was chosen as a fallback site.
However, this was a weird place for a neighborhood, and could
sometimes cause problems with building roads. There was also no
guarantee that the corner point wasn't covered in water.

Now when a valid neighborhood site doesn't exist, neighborhood
generation will stop and return the set of neighborhoods sites
selected so  far (which might be none). This seems like better and
more expected behavior.

As a side effect, some of the logic in `WorldGenerator.generate()`
needs to be updated to handle the case where there is an empty set
of neighborhood sites even though more than zero sites were
desired, which previously couldn't happen.