Adding explicit `browserslist` config
This adds support for a larger set of older browsers than the
default set of browsers used when no configuration is given. The
list of browsers should also now be more stable over time, since
the new config should not automatically drop older browser
versions as new versions are released.

For production builds, the generated JavaScript and CSS output
should be identical to the output generated before this change was
made.

While doing this, also upgrading the `browserslist` and
`caniuse-lite` packages to the most recent versions, to make sure
the browser info is up to date.