Upgrading Webpack v4 -> v5
Also upgrading relevant Webpack plugin packages to the latest
versions.

`optimize-css-assets-webpack-plugin` is being removed because it
apparently doesn't support Webpack 5. Its docs say to use
`css-minimizer-webpack-plugin` instead.

Also, with Webpack 5 having two declarations of `TerserPlugin` in
`webpack.config.js` no longer seems to cause the generated
JavaScript output to be smaller, so the duplicate declaration is
being removed.

The remaining `TerserPlugin` configuration is being changed so
that the existing Webpack 4 behavior of not minifying JavaScript
files in the `lib/` folder is preserved.