Converting three.js dependency to npm package
First the downside: on my laptop this change appears to slow down
production builds by ~1.9x, and development builds by ~1.1x.

Development: 3.901sec ->  4.264sec
Production:  5.734sec -> 10.918sec

(Average of 10 builds).

However, this change makes it easier to upgrade three.js by
removing the need to manually download a minified *.js file and
copy it into the repository. Instead, it can now be upgraded via
the `package.json` file in the same way as the other dependencies.

It also means the `THREE` namespace can be selectively imported
via `import` statements, instead of there being a `THREE` object
defined as a top-level global variable.

Finally, this reduces the combined download size of the app's
JavaScript code by 53.4KB (un-GZIPed), or 19.5KB (GZIPed)
according to Safari's web inspector. This is presumably due to
three.js's improved support in recent versions for tree shaking
when an app is built using Webpack.

Un-GZIPed
=============================
*.js code: 700.6KB -> 647.2KB
Full App:  709.8KB -> 656.3KB

GZIPed
=============================
*.js code: 181.1KB -> 161.6KB
Full App:  185.1KB -> 165.5KB

Hopefully the downside of slower builds will not be too much of a
pain. If it is, then this change can be reverted in the future.