Buildings can be built along west/north edge roads
Previously, buildings could be built on land opposite the
easternmost or southernmost edge of the road network, but not the
westernmost or northernmost edge.

This was caused by an incorrect bounding box being used to
indicate where buildings could be built. The logic previously used
the raw bounding box of the road network. But since blocks of
buildings are indexed by their northwestern corner, this caused
land immediately north of the northern edge or west of the western
edge of the road network to be excluded.