Buildings on edges of road network are rendered
Previously, buildings on the outside side of a road network
bounding box edge would not be rendered. For example, if a
north/south road was the westernmost road in the city, any
buildings on the western side of the road would not be rendered.

This was caused by an off-by-one error in each cardinal direction
when adding buildings to the scene.