Instanced buildings include antennas
The brings the version of building mesh generation that uses
`THREE.BufferGeometry` to parity with the legacy version that uses
`THREE.Geometry`.

The new `ANTENNA_WIDTH` and `ANTENNA_DEPTH` constants are set with
values that should make antenna sizes the same as in the legacy
implementation. In the legacy implementation each antenna is built
using a cylinder configured with a given radius. However, since
the cylinder only has 4 sides in practice it is a box. I believe
this means the "radius" is half the diagonal of the square top and
bottom faces.

Since the top and bottom cylinder faces are each composed of two
right triangles where the two non-hypotenuse sides are equal, the
side lengths (i.e. width and depth) can be calculated by
multiplying the diagonal (i.e. hypotenuse) by sqrt(2) / 2.

More explicitly:

Cylinder radius = 0.016666666666667
Diagonal/Hypotenuse = 0.016666666666667 * 2 = 0.033333333333334
Sides = 0.033333333333334 * (sqrt(2) / 2) = 0.023570226039552

Therefore, 0.023570226039552 is used as the value for
`ANTENNA_WIDTH` and `ANTENNA_DEPTH`.