Converting road meshes to `THREE.BufferGeometry`
Instead of `THREE.Geometry`, which is not present in three.js v125
and newer. Upgrading past v124 requires removing all usage of
`THREE.Geometry` in the app, and this is a step in that direction.

Another benefit is that this new implementation creates the road
meshes faster in Safari, Firefox, and Chrome. The numbers below
are the average road mesh generation times of 10 random cities
created with the default settings using a production build on my
laptop:

Safari 15.6.1
--------------------------------
Before:  396.3ms
After:    24.3ms
Delta:   372.0ms (16.31x) faster

Firefox 104.0.1
--------------------------------
Before:  256.7ms
After:    22.9ms
Delta:   233.8ms (11.21x) faster

Chrome 105.0.5195.52
--------------------------------
Before:  277.6ms
After:    29.9ms
Delta:   247.7ms  (9.28x) faster

As part of this, some values on the `Config` object are either
removed or moved to `RoadMeshBuilder`, because they are either no
longer needed or are specific to generating the road meshes and
don't need to be exposed globally.