Converting debug mesh to `THREE.BufferGeometry`
Instead of `THREE.Geometry`, which is not present in three.js v125
and newer. This commit removes the final usage of `THREE.Geometry`
from the app, so three.js should now be upgradable past v124.

The original implementation was already fast, but like the other
mesh generators this seems to gain a speed boost from the switch
to `THREE.BufferGeometry`. From a few informal tests, the old
version seemed to complete in 1-3ms on my laptop in Safari,
Firefox, and Chrome, while the new version completes in 0-1ms. (I
think 0ms means "less than 1ms").