Fixing grid background for three.js v137+
This commit defines the background grid material in a different
way that should allow it to still have the same appearance after
three.js is upgraded in the future to v137 or later.

Starting in v137, materials that use a texture with transparent
pixels must be explicitly marked as transparent. Otherwise, any
transparent pixels will be rendered as opaque; when using the
existing `textures/grid.png` image, the transparent areas between
the background grid lines will be rendered as solid black.

Manually configuring the material to be transparent removes the
solid black pixels. However, the blending equation also needs to
be changed to use the equation for premultiplied alpha blending.
Otherwise, each of the grid lines will have a black halo, causing
them to appear thicker and darker than intended.

I'm not sure why when using v124 the grid background looks the
same with both the old and new material definitions. Maybe when
the material was not marked as transparent but the texture had
transparency v124 was internally doing the equivalent of the new
configuration? I'm not really sure though.