Reducing scope of variables
These variables are only used in `build()`, except for in
`addQuad()` where they are passed in as an argument. Therefore,
they can be defined in `build()` instead of a broader scope.

As a result of this change, they no longer need to be nulled out
at the end of `build()` since they should now be garbage collected
automatically.