More accurate constant name
The previous name was inaccurate because the surface height of a
terrain point with water will not necessarily be 0; it will
actually typically be above 0 (but I believe never below 0). The
new name more directly indicates what the constant represents in
practice. It also provides some future-proofing if bridges can one
day be built over land.

As part of the name change the definition is being moved to be
next to the other bridge support column constant definitions.