Better units for specifying border radius
Each slider thumb is intended to be a circle, and setting the
border radius to 50% is a simple way to make this so when the
width and height are the same. Although the previous `rem` units
gave the same result, the new units should be more future-proof
because they shouldn't need to change if the width and height
change in the future.