Removing `-webkit` Flexbox fallback styles
Chrome has supported unprefixed Flexbox styles since v29 (released
on July 17, 2013 according to caniuse.com), and Safari since v9
(released on September 30, 2015). For these fallback styles to be
needed a visitor would presumably need to be using a browser
version more than 7 years old as of this commit. Based on that,
it seems safe to remove them.