Fixing potential memory leaks
`NavigationTouchController`, `TimerLoop`, and `MapCamera` are
designed to exist for the lifetime of the page, but with some
configuration changed each time a new city is created. The
`initialTerrain` and `initialWorldData` variables were set at
object creation, and then existed for the relevant object's
lifetime, even after new terrain or world data was set. This had
the potential to be a memory leak.

Why only a potential? The reason is that the initial terrain or
world data objects that were used intentionally also exist in
`App` for the lifetime of the page, so I don't think any memory
was actually leaked. These objects also are for empty terrain or
world data, and therefore use little memory anyway. However, it
still seems like a good idea to make this change to prevent
potential issues in the future.