Fixing "generate buildings?" flag
Before this commit, if `GENERATE_BUILDINGS` was set to `false` in
`WorldGenerator`, then `generate()` would fail with an error.

I believe this bug was introduced in commit
eeda3790dd1bf1262d13fcf8f2c2aaa2438b403b. In that commit the
output format from `ZonedBlockGenerator.generate()` changed, but
this logic wasn't updated to reflect that.

To try to reduce the chance of this bug reoccurring, the mechanism
of generating an empty set of buildings is being changed. Rather
than hard-coding an empty zoned block object whose format might
need to change in the future, zoned blocks are instead generated
using `ZonedBlockGenerator.generate()`, but using an empty road
network so that there is nowhere to place any buildings.