Fixing `removeSubscriber()` return value
Previously, `MessageBroker.removeSubscriber()` would correctly
remove a valid subscription, or do nothing if no subscription with
the given ID existed. However, the return value from the function
was always `false`, instead of `true` if a subscription was
removed and `false` otherwise.

The source of the bug was that the `return true` statement upon a
successful subscription removal only returned from the inner
`forEach()` function (and was essentially a no-op), rather than
from the parent function.

A side benefit of this new implementation is that it will also
exit early when the relevant subscription is found, unlike the
previous implementation. This shouldn't change semantics because
the implementation should ensure that only one subscription is
ever tied to a particular ID.