Removing toggling of unneeded CSS class
The presence or absence of this class didn't make any difference
in the appearance/behavior of the menus. I believe this CSS class
was a no-op; since the parent `"menus-container"` element is a
block flex container any of its flex items with `inline-block`
display should be treated as block elements instead.