Gestures can be performed while a menu is open
Previously, if the "New City" or "About" menu was open, a "click"
event on the main canvas would close the menu, but other gestures
(pan/zoom/rotate/tilt) would not work. Now, a "mousedown" or
"touchstart" event on the main canvas will immediately close any
open menu and allow gestures to be performed like normal.