Adding "Dev" menu as a new home for debug toggles
The "Dev" menu is not normally present, but can be added to the
page by double-clicking/tapping the "City Tour" label in the
"About" menu.

This menu is intended as a way to expose runtime toggles for debug
tools, which should allow more flexibility than constant flags
that require a recompile to change.

This commit only implements an empty placeholder menu. Populating
the menu will happen in future commits.