Adding toggle to show neighborhood centers
Previously, whether or not to show the debug markers for each
neighborhood center could be configured with a hard-coded flag,
but the value couldn't be changed at runtime. The new toggle lives
in the "Dev" menu.

In order to allow the toggle to take immediate effect without the
camera having to be moved first, a new `makeDirty()` function that
forces a re-render on the next timer tick is being added to the
object returned by `RenderView()`.