Adding UI debug toggle to show driving path
A visualization of the path the camera will take during the
"driving" portion of a flythrough can now be added to the scene,
via a toggle in the "Dev" menu. It is updated whenever a driving
path is generated, and persists even after a flythrough stops.
It is intended to be an aid when working on this portion of the
flythrough. Previously, the pieces needed to display this
visualizaton were in place, but required adding a small amount of
code and rebuilding the app.

The UI refers to this as the "driving path", since in practice
that is what it is. However, the code refers to "route curves", as
a way to future-proof against the possibility of showing flight
routes as well.

Since building the route curves mesh sometimes takes a large
enough time (up to around 10ms on my laptop), it is only added to
the scene when the toggle is enabled. (Instead of always adding it
to the scene, and only toggling the visibility).