Fixing initial flythrough driving segment duration
Previously, the initial driving segment of the flythrough lasted
twice as long as intended. The reason was that the "initial
descent" portion of the flythough contains a full driving
segment, and was followed by an additional regular driving
segment.

The fix in this commit is being done for expediency. It would be
better to implement this by changing the "initial descent" portion
to not include a driving segment. However, it seems like that will
require a more involved change to implement properly. This commit
fixes the behavior in the meantime.