Renaming vehicle mode "initial descent" -> "intro"
"Intro" seems like a better general purpose (and future proof)
name. "Initial descent" was potentially confusing, because this
mode includes a rise above the city center, in addition to the
subsequent descent to street level.

As a side benefit, this adds a "_MODE" suffix to the relevant
constant name for consistency with the other constants
representing vehicle modes.