Fixing vehicle Y coordinate calculation
When setting the Y coordinate, `roadHeightAtCurrentPosition()` is
used to prevent the camera from going below the terrain, and this
function in turn uses the vehicle's current X and Z coordinates.
However, previously only the vehicle's X coordinate was updated
before calling this function, so it was using the current frame's
X coordinate but the previous frame's Z coordinate.

I'm not sure whether the old behavior caused much of a practical
problem, but the calculation should now be correct.