Preventing camera jumps during driving U-turn
Previously, the camera would sometimes suddenly jump to a new
angle when performing a U-turn, instead of rotating 180 degrees in
a continuous motion. The problem was caused by using the incorrect
angle for the start of the rotation. Instead of using the ending
angle of the previous driving segment, it used the starting angle.