Reordering logic to improve readability
This organizes the logic into two phases: first, the data needed
for the flythrough intro animations is calculated, and then the
animations are built using that data. Further, the logic to
calculate the animation data is now in the same order that the given
animations occur. Previously, this logic was spread around, and
subjectively it made the code a little harder to understand.