Better terminology for flythrough intro animations
The flythrough intro animations consist of traveling to a point
above the city, diving down to street level, and then driving.

Previously, "target" was used to refer to the point above the city,
and "descentTarget" was used to refer to the point at street level
that the dive went to. However, "target" was non-descriptive and
potentially confusing, and "dive" seems like a more descriptive
term than "descent".

Now these two points are referred to as "topOfDive" and
"bottomOfDive" respectively, which seems more explicit and clear.

The relevant variable declarations are also being converted from
`var` to `let` for code style.