More generic terminology in flythrough intro code
This removes references to "city center" in variable names and
comments in the code that implements the flythrough intro
animation. Although the flythrough is currently configured to
always start by flying to the city center, the code is intended to
be flexible and not make assumptions about that.

Relevant variable declarations are also being converted from `var`
to `let` for code style.