Better variable name, terms in angle calculation
Although equivalent, it seems more clear to refer to the relevant
angle as the rotation angle at the bottom of the dive, rather than
the rotation angle at the start of driving. This allows the dive
animation to be defined in terms of the top of the dive and the
bottom of the dive.

Similarly, it seems more clear to calculate this angle using the
X/Z coordinates at the bottom of the dive, rather than the ending
X/Z coordinates of the initial 3 block portion of the driving
segment. Both options should result in the same angle, since the 3
block driving portion always occurs in the same direction as the
dive.