Preventing confusing rotation in flythrough intro
The flythrough intro animation attempts to dive down to a point at
street level 3 blocks away from the top of the dive. If there
isn't a road at that position, as a fallback it will dive straight
down instead.

Previously in this scenario the camera would always rotate to
face north (i.e. 0 degrees) at the bottom of the dive, regardless
of which way the camera was facing at the top of the dive. This
could be confusing and disorienting. This happened because
`azimuthAngleToPoint()` returns 0 when the start and target
coordinates are the same.

Now in the fallback scenario the camera will instead rotate toward
the cardinal direction closest to the camera's azimuth angle at
the top of the dive. Although there is probably an even better
animation that could be used, this seems like an improvement over
the previous behavior.