Forward drive always attempted after intro dive
During the flythrough intro animation, an attempt is made to dive
down to street level 3 blocks away from the top of dive, and then
drive 3 more blocks in the same direction before performing any
turns. However, if the point 3 blocks away from the top of the
dive does not have a road intersection then the dive will instead
happen straight down.

Previously in this situation no attempt would be made to drive
forward in the same direction either. This would typically mean
that a turn would happen immediately upon the end of the dive as
the driving segment began.

However, there is sometimes room to drive 1 or 2 blocks forward in
this situation. After this commit that will now occur if that is
the case. This results in a smoother post-dive animation.