Preventing >180 degree turns in flythrough
Previously the camera would sometimes make a turn in airplane or
helicopter mode by turning 270 degrees, instead of turning 90
degrees in the opposite direction. In certain situations it was
also possible for an overly large turn to occur in the "fly to
city center" portion of the intro animation.

During the flythrough the camera's Y-axis rotation can increase or
decreased unboundedly as a result of turning in the same direction
many times in a row (more likely to occur in a small city).
However, target azimuth angles are always initially in the range
0-2π. As a result the target angle sometimes needs to be corrected
by more than 2π to prevent a turn larger than 180 degrees.
However, the correction logic previously only tried to correct the
angle by at most 2π, resulting in an overly large turn.