Removing unneeded `RoadNavigator` variables
Previously, `subTargetX` and `subTargetZ` stored the current step
of the driving path, while `targetX` and `targetZ` stored the
final step. However, dedicated variables to store the final step
are redundant because that data can be obtained from the path
itself, and `targetX` and `targetZ` weren't actually used for
anything.

After this commit `targetX` and `targetZ` now refer to the current
step on the path, and `subTargetX` and `subTargetZ` are removed.
The resulting code seems more clear than before.