Preventing `RoadNavigator.nextTarget()` error
Previously, `RoadNavigator.nextTarget()` could sometimes throw an
error during the flythrough. The current position was sometimes
chosen as the next target point to drive to, resulting in an empty
driving path, which would cause an error. Although unlikely to
happen in a larger city, this seemed to happen occasionally in a
small cities (such as with only 1 neighborhood).

This commit fixes the logic that chooses the next target point so
that it never chooses the current position, which should prevent
this error from occurring.