Preventing error when generating road network
Previously an error could sometimes occur during city generation
due to attempting to build a road outside of the valid bounds of
the road network. This would cause the "Building new city..."
overlay to be shown forever, requiring a page refresh to use the
app again.

The culprit was `TerrainCandidateRoadNetwork.edgesFrom()`. When
returning the edges from a point it could include edges outside
the valid bounds of the road network. This could cause the chosen
path from a neighborhood center to the rest of the road network to
include invalid out of bounds points. When this path was converted
into actual roads the error would occur.

This error was more likely to occur when building a large city
with hilly terrain, because that would make it more likely a
neighborhood center was near the edge of the terrain and had
constrained options for connecting to the rest of the road
network.

Although `TerrainCandidateRoadNetwork.edgesFrom()` would correctly
only return edges that were inside the bounds of the terrain
(which has larger bounds than the road network), this was by
accident. When an edge was outside of the bounds certain values
would be `undefined` instead of a number, which accidentally
caused the "is this edge valid" conditions to evaluate to `false`.
A side benefit of this commit is that the correct behavior will
now instead happen more intentionally.

This bug appears to have been introduced in commit
26e9ecfed4960273be2819b3c621d30ac7c7ae9a. That commit should have
included the changes in this commit, but they were overlooked.
Although, you could argue they should have been added even earlier
to fix the "accidentally correct" behavior mentioned above.