Upgrading three.js v151.3 -> v152.2
This includes configuration changes to use the defaults from v151,
instead of the new defaults introduced in v152. This prevents the
scene's appearance from changing due to sRGB being the default
output color space in v152, instead of linear sRGB. It might be
good to revisit things in the future so that sRGB can be used as
the output color space, but this commit allows using the most
recent version of three.js in the meantime.