Fixing initial config values to be in sync
Previously the two affected fields had different initial config
values in the JavaScript code and the HTML markup. It looks like
this has been the case ever since the scene dimensions were
rescaled in commit f43e94b7f4c5b0540b3addb282ad219a47db3b1b. I
assume the reason for the difference was that a repeating decimal
doesn't work well as a slider value?

It seems better for the JavaScript and HTML values to always be in
sync. This change errs on the side of using the HTML values, since
the UI sliders have less flexibility in regard to the values that
can be set. I don't notice any obvious differences between cities
generated with the old values and new values.