Moving terrain height controls to "Dev" menu
The ability to generate terrain with mountains and hills instead
of a mostly flat expanse is not ready yet. Moving the controls
that allow addings hills out of the main "New City" menu to the
hidden "Dev" menu is a step toward preventing this functionality
from being a blocker for deploying updated versions of the app.

A side benefit of this change is that the relevant sliders will
now have their initial values automatically set, rather than
having a hard-coded initial value.

Another side benefit is that this change will make it easier to
use the "Jitter" and "Ruggedness" sliders during development,
since it will no longer require manually editing the HTML markup
in the running app to make the controls visible.