Configuring cities to not have hills by default
The ability to generate terrain with mountains and hills instead
of a mostly flat expanse is not ready yet. To prevent this
functionality from being a blocker for deploying updated versions
of the app, cities are now configured to not have any hills by
default. The relevant config values can still be changed, but it
requires using the hidden "Dev" menu.

To prevent the terrain from being _too_ flat and boring, the
default "Jitter" and "Ruggedness" settings are being changed so
that the terrain will be slightly gently rolling.