Disabling unneeded `webpack-dev-server` features
`--no-bonjour`: I believe Bonjour should already be disabled by
default, but it is being explicitly disabled to be defensive in
case that changes in the future.

`--no-static`: The app doesn't have any assets (static or
otherwise) that are served outside of the `dist/` folder.
Therefore there doesn't seem to be any reason to enable the
ability to also serve assets from an additional folder.