Setting specific width for slider elements
Previously slider elements didn't have a specific width set, which
led to the browser picking a width instead. Safari (desktop and
mobile) and Chrome used a width of 129px, while Firefox used a
width of 160px.

Now all sliders are set to have a width of 128px. This value is
very similar to the previous Safari and Chrome value, but has the
benefit of being a multiple of the root font size of 16px.