Converting nav. controls toggle `<a>`->`<button>`
Semantically this element seems like a button, not a hyperlink.
After this change the `no-underline` CSS class is no longer used
anywhere, so it is being removed.

Commit 667bf696353393a35404503f63d86899de95e47e made similar
changes to menu title elements. As in that commit, a minor change
is that the mouse cursor over this element is now the default
mouse cursor, instead of the "pointing hand" cursor. (At least on
macOS, not sure about other OSes). Although the "pointing hand"
cursor _could_ be preserved, the default cursor seems more
consistent with the normal behavior of buttons, and the hover
appearance for this element already provides an affordance that it
is clickable.

Unlike that commit, no effort is being made to prevent double taps
on mobile to prevent page zooming. I don't think it's possible to
double tap this element, because it immediately moves away once
you tap it once.